requirements.txt
opencv-python
face-recognition
mysql-connector-python
Pillow
numpy
